+++ machine/plugins/builtin/debug.py
@@ -1,10 +1,11 @@
-import logging
 from typing import Any
+
+from structlog.stdlib import get_logger
 
 from machine.plugins.base import MachineBasePlugin
 from machine.plugins.decorators import process
 
-logger = logging.getLogger(__name__)
+logger = get_logger(__name__)
 
 
 class EchoPlugin(MachineBasePlugin):
@@ -15,8 +16,5 @@ class EchoPlugin(MachineBasePlugin):
             "subtype" in event and (event["subtype"] == "bot_message" or event["subtype"] == "message_replied")
         ) or event["user"] == self.bot_info["user_id"]:
             return
-        if "thread_ts" in event:
-            thread_ts = event["thread_ts"]
-        else:
-            thread_ts = None
+        thread_ts = event.get("thread_ts")
         await self.say(event["channel"], event["text"], thread_ts=thread_ts)