+++ CHANGELOG.md
@@ -8,6 +8,107 @@ _prefixes_. This project adheres to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 
+## [0.37.0] - 2024-05-26
+
+### Added
+
+- Support for **Block actions**, coming from interactive elements from [Block Kit](https://api.slack.com/block-kit)
+  ([#1034](https://github.com/DonDebonair/slack-machine/pull/1034) with help from
+  [**pawelros**](https://github.com/pawelros))
+
+### Changed
+
+- Bump `slack-sdk` from 3.27.1 to 3.27.2
+
+## [0.36.0] - 2024-05-04
+
+### Added
+
+- Replace `flake8`, `isort` and `black` with `ruff` ([#1027](https://github.com/DonDebonair/slack-machine/pull/1027))
+- Add Python 3.12 ([#1028](https://github.com/DonDebonair/slack-machine/pull/1028))
+
+### Changed
+
+- Update Github Actions ([#1008](https://github.com/DonDebonair/slack-machine/pull/1008))
+- Bump `slack-sdk` from 3.21.3 to 3.27.1
+- Bump `httpx` from 0.24.1 to 0.27.0
+- Bump `pydantic` from 2.3.0 to 2.7.1
+- Bump `cryptography` from 41.0.3 to 42.0.4
+- Bump `aiohttp` from 3.8.5 to 3.9.5
+- Bump `urllib3` from 1.26.16 1.26.18
+- Bump `structlog` from 23.1.0 to 24.1.0
+- Bump `redis` from 5.0.0 to 5.0.4
+- Bump `pyee` from 11.0.0 to 11.1.0
+- Bump `aioboto3` from 11.3.0 to 12.4.0
+- Bump `tzdata` from 2023.3 to 2024.1
+- Bump `aiosqlite` from 0.19.0 to 0.20.0
+- Bump `hiredis` from 2.2.3 to 2.3.2
+
+## [0.35.0] - 2023-09-03
+
+### Added
+
+- Incoming requests/events from Slack that the Slack App is subscribed to, will be logged when `LOGLEVEL` is set to
+  `DEBUG` ([#876](https://github.com/DonDebonair/slack-machine/pull/876))
+
+### Changed
+
+- **Breaking:** the optional `init()` method of plugins is now expected to be `async`. This allows plugin authors to
+  interact with Slack during plugin initialization through Slack Machine's plugin API.
+  ([#868](https://github.com/DonDebonair/slack-machine/pull/868))
+- Standard app manifest for Slack Machine now also enables listening for the `app_home_opened` event
+- Bump `aiosqlite` from 0.18.0 to 0.19.0
+- Bump `apscheduler` from 3.10.1 to 3.10.4
+- Bump `redis` from 4.6.0 to 5.0.0
+- Bump `pyee` from 10.0.2 to 11.0.0
+- Bump `pydantic` from 2.2.1 to 2.3.0
+
+### Fixed
+
+- Use conversations_setTopic instead of channels_setTopic for setting channel topic
+  ([#869](https://github.com/DonDebonair/slack-machine/pull/869) by
+  [**@jogendra**](https://github.com/jogendra))
+
+## [0.34.2] - 2023-08-13
+
+### Fixed
+
+- Users indexed by email and related functions are now exposed through the plugin interface
+  ([#852](https://github.com/DonDebonair/slack-machine/pull/852))
+
+## [0.34.1] - 2023-08-13
+
+### Added
+
+- Slack users are now indexed by their email as well, allowing fast lookups by email
+  ([#849](https://github.com/DonDebonair/slack-machine/pull/849))
+
+### Fixed
+
+- All logging in Slack Machine is now done through structlog instead of the Python stdlib logger
+  ([#850](https://github.com/DonDebonair/slack-machine/pull/850))
+
+## [0.34.0] - 2023-08-13
+
+### Added
+
+- Add support to set topic on channels ([#839](https://github.com/DonDebonair/slack-machine/pull/839) by
+  [**@jogendra**](https://github.com/jogendra))
+- Add SQLite storage backend ([#844](https://github.com/DonDebonair/slack-machine/pull/844) by
+  [**@cp-richard**](https://github.com/cp-richard))
+
+### Changed
+
+- Bump `httpx` from 0.24.0 to 0.24.1
+- Bump `aiohttp` from 3.8.4 to 3.8.5
+- Bump `certifi` from 2022.12.7 to 2023.7.22
+- Bump `pyee` from 9.1.0 to 10.0.2
+- Bump `pydantic` from 1.10.7 to 2.1.1 ([#840](https://github.com/DonDebonair/slack-machine/pull/840))
+
+### Removed
+
+- **Breaking:** Remove Python 3.7 support ([#846](https://github.com/DonDebonair/slack-machine/pull/846))
+
 ## [0.33.0] - 2023-05-15
 
 ### Added
@@ -576,7 +677,13 @@ _First release. Rejoice!_ :wave:
 - A simple, yet powerful and extendable Slack bot framework
 
 
-[Unreleased]: https://github.com/DonDebonair/slack-machine/compare/v0.33.0...HEAD
+[Unreleased]: https://github.com/DonDebonair/slack-machine/compare/v0.37.0...HEAD
+[0.37.0]: https://github.com/DonDebonair/slack-machine/compare/v0.36.0...v0.37.0
+[0.36.0]: https://github.com/DonDebonair/slack-machine/compare/v0.35.0...v0.36.0
+[0.35.0]: https://github.com/DonDebonair/slack-machine/compare/v0.34.2...v0.35.0
+[0.34.2]: https://github.com/DonDebonair/slack-machine/compare/v0.34.1...v0.34.2
+[0.34.1]: https://github.com/DonDebonair/slack-machine/compare/v0.34.0...v0.34.1
+[0.34.0]: https://github.com/DonDebonair/slack-machine/compare/v0.33.0...v0.34.0
 [0.33.0]: https://github.com/DonDebonair/slack-machine/compare/v0.32.0...v0.33.0
 [0.32.0]: https://github.com/DonDebonair/slack-machine/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/DonDebonair/slack-machine/compare/v0.30.0...v0.31.0